+++ src/app/admin/users/page.tsx
@@ -1,6 +1,6 @@
-import CreateUsers from "@/components/admin/CreateUsers";
-import OnlyRoles from "@/components/OnlyRoles";
+import OnlyRoles from "@/components/auth/OnlyRoles";
 import PageWithHeader from "@/components/PageWithHeader";
+import UsersDataManager from "@/components/admin/users/UsersDataManager";
 import React from "react";
 
 export const metadata = {
@@ -11,7 +11,7 @@ export default function Users() {
     return (
         <OnlyRoles roles={["administrator"]}>
             <PageWithHeader title="Admin / Felhasználók" homeLocation="/admin">
-                <CreateUsers />
+                <UsersDataManager />
             </PageWithHeader>
         </OnlyRoles>
     );